+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The template version should stay in sync with the engine version
 - Twinejs shouldn't save to localstorage at all
 - No longer require a dummy start passage
 - Bugfix twine passage tags
+- `host.tiles.toggle()` for toggling sprites
 
 ## 4.0.0 - 7/22/25